@@ -1,31 +1,39 @@
 {
-    "name": "slim/flash",
+    "name": "firehead996/flash-messages",
     "type": "library",
-    "description": "Slim Framework Flash message service provider",
-    "keywords": ["slim","framework","provider","flash","message"],
-    "homepage": "http://slimframework.com",
+    "description": "Flash message service provider",
     "license": "MIT",
     "authors": [
         {
             "name": "Josh Lockhart",
             "email": "hello@joshlockhart.com",
             "homepage": "http://joshlockhart.com"
+        },
+        {
+            "name": "Mikolaj Panecki",
+            "email": "firehead996@gmail.com"
         }
     ],
     "require": {
-        "php": ">=5.5.0"
+        "php": "^8.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.0"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
-            "Slim\\Flash\\": "src"
+            "Firehead996\\Flash\\": "src"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Slim\\Flash\\Tests\\": "tests"
+            "Firehead996\\Flash\\Tests\\": "tests"
         }
+    },
+    "scripts": {
+        "test": [
+            "@phpunit"
+        ],
+        "phpunit": "phpunit"
     }
 }
